#  middleware: Preallocate middleware ordering slices to reduce allocations