Updates the middleware ordering behavior to preallocate its order item slices/maps to prevent future allocations.

Fixes #162

Example of v2 SDK's benchmark different

```go
benchmark                                          old ns/op     new ns/op     delta
BenchmarkCustomizations/small/smithy/default-4     98471         80525         -18.22%
BenchmarkCustomizations/small/smithy/default-4     127331        95931         -24.66%
BenchmarkCustomizations/small/smithy/default-4     143848        87465         -39.20%
BenchmarkCustomizations/small/smithy/default-4     250105        98629         -60.56%
BenchmarkCustomizations/small/smithy/default-4     190916        93149         -51.21%
BenchmarkCustomizations/small/smithy/default-4     149556        86324         -42.28%
BenchmarkCustomizations/small/smithy/default-4     171215        86332         -49.58%
BenchmarkCustomizations/small/smithy/default-4     149547        84465         -43.52%
BenchmarkCustomizations/small/smithy/default-4     138599        95963         -30.76%
BenchmarkCustomizations/small/smithy/default-4     145238        100273        -30.96%

benchmark                                          old allocs     new allocs     delta
BenchmarkCustomizations/small/smithy/default-4     681            675            -0.88%
BenchmarkCustomizations/small/smithy/default-4     681            675            -0.88%
BenchmarkCustomizations/small/smithy/default-4     681            675            -0.88%
BenchmarkCustomizations/small/smithy/default-4     681            675            -0.88%
BenchmarkCustomizations/small/smithy/default-4     681            675            -0.88%
BenchmarkCustomizations/small/smithy/default-4     681            675            -0.88%
BenchmarkCustomizations/small/smithy/default-4     681            675            -0.88%
BenchmarkCustomizations/small/smithy/default-4     681            675            -0.88%
BenchmarkCustomizations/small/smithy/default-4     681            675            -0.88%
BenchmarkCustomizations/small/smithy/default-4     681            675            -0.88%

benchmark                                          old bytes     new bytes     delta
BenchmarkCustomizations/small/smithy/default-4     51196         51068         -0.25%
BenchmarkCustomizations/small/smithy/default-4     51197         51069         -0.25%
BenchmarkCustomizations/small/smithy/default-4     51197         51069         -0.25%
BenchmarkCustomizations/small/smithy/default-4     51197         51069         -0.25%
BenchmarkCustomizations/small/smithy/default-4     51198         51068         -0.25%
BenchmarkCustomizations/small/smithy/default-4     51197         51068         -0.25%
BenchmarkCustomizations/small/smithy/default-4     51197         51068         -0.25%
BenchmarkCustomizations/small/smithy/default-4     51197         51068         -0.25%
BenchmarkCustomizations/small/smithy/default-4     51197         51068         -0.25%
BenchmarkCustomizations/small/smithy/default-4     51197         51068         -0.25%
```